# Conflicts:
#	src/solver/optimisation/CMakeLists.txt
#	src/solver/optimisation/LegacyNameMapper.cpp
#	src/solver/optimisation/include/antares/solver/optimisation/LegacyNameMapper.h
#	src/solver/optimisation/opt_appel_solveur_lineaire.cpp
#	src/tests/cucumber/features/solver-features/legacy_simulation_table.feature
#	src/tests/src/solver/optimisation/CMakeLists.txt
#	src/tests/src/solver/optimisation/test_legacy_name_mapper.cpp

+++ src/libs/antares/study/parts/hydro/container.cpp
@@ -706,7 +706,7 @@ std::optional<ReserveID> PartHydro::reserveParticipationAt(const Area* area,
     {
         if (reserveParticipationContainer.value().isParticipatingInReserve(reserveID))
         {
-            if (globalReserveParticipationIdx == index)
+            if (static_cast<unsigned int>(globalReserveParticipationIdx) == index)
             {
                 return reserveID;
             }
